refactor(web): insert graphql queries/mutations/subscriptions to dedicated files
