+++ apps/web/src/features/chat/hooks/use-recent-chats-subscription.ts
@@ -1,61 +1,21 @@
 import { useSubscription } from "urql";
 
 import { useRecentChatsQuery } from "@/features/chat/hooks/use-recent-chats-query";
-import { graphql } from "@/gql";
-
-const RecentChatsSubscription = graphql(`
-  subscription ChatUpdated {
-    chatUpdated {
-      id
-      title
-      participants {
-        firstName
-        lastName
-        id
-      }
-      latestMessage {
-        id
-        sender {
-          id
-          firstName
-        }
-        content
-        createdAt
-      }
-    }
-  }
-`);
-
-type ChatPreview = {
-  id: string;
-  title?: string | null;
-  participants: {
-    firstName: string;
-    lastName: string;
-    id: string;
-  }[];
-  latestMessage: {
-    id: string;
-    sender: {
-      id: string;
-      firstName: string;
-    };
-    content: string;
-    createdAt: string;
-  };
-};
+import type { ChatUpdatedSubscription } from "@/gql/graphql";
+import { RECENT_CHATS_SUBSCRIPTION } from "@/lib/gql/subscriptions";
 
 export const useRecentChatsSubscription = () => {
   const [queryResult] = useRecentChatsQuery();
 
   const [subResult] = useSubscription(
     {
-      query: RecentChatsSubscription,
+      query: RECENT_CHATS_SUBSCRIPTION,
       pause: queryResult.fetching,
     },
     (
-      chats: ChatPreview[] = queryResult.data?.recentChats ?? [],
-      response: { chatUpdated: ChatPreview }
+      chats: ChatUpdatedSubscription["chatUpdated"][] = queryResult.data?.recentChats ??
+        [],
+      response: ChatUpdatedSubscription
     ) => {
       const chatWithoutUpdated = chats.filter((el) => el.id !== response.chatUpdated.id);
       return [response.chatUpdated, ...chatWithoutUpdated];
